requirements.txt
asgiref==3.8.1
Django==5.2.1
django-cors-headers==4.7.0
django-debug-toolbar==4.4.2
djangorestframework==3.16.0
python-decouple==3.8
sqlparse==0.5.3
